Moving out node api functions into a separate file.
