+++ engine/src/core/scripting/python/pocketpy/api/cre_pkpy_api.h
@@ -1,0 +1,5 @@
+#pragma once
+
+struct pkpy_vm_handle;
+
+void cre_pkpy_api_load_internal_modules(struct pkpy_vm_handle* vm);
