+++ engine/src/core/scripting/python/pocketpy/api/cre_pkpy_api_node.c
@@ -1,632 +1,26 @@
-#include "cre_pkpy_api.h"
+#include "cre_pkpy_api_node.h"
 
 #include <string.h>
 
-#include <seika/rendering/frame_buffer.h>
+#include <pocketpy_c.h>
+
 #include <seika/rendering/renderer.h>
 #include <seika/rendering/shader/shader_cache.h>
 #include <seika/asset/asset_manager.h>
+#include <seika/utils/logger.h>
 #include <seika/utils/se_string_util.h>
 #include <seika/utils/se_assert.h>
 
-#include "cre_pkpy.h"
-#include "cre_pkpy_util.h"
-#include "cre_pkpy_api_source.h"
-#include "../../../ecs/entity/entity.h"
-#include "../../../ecs/component/component.h"
-#include "../../../ecs/component/node_component.h"
-#include "../../../ecs/component/sprite_component.h"
-#include "../../../ecs/component/animated_sprite_component.h"
-#include "../../../ecs/component/text_label_component.h"
-#include "../../../ecs/component/collider2d_component.h"
-#include "../../../ecs/component/color_rect_component.h"
-#include "../../../ecs/component/parallax_component.h"
-#include "../../../scene/scene_utils.h"
-#include "../../../ecs/component/script_component.h"
-#include "cre_pkpy_entity_instance_cache.h"
+#include "../cre_pkpy_entity_instance_cache.h"
+#include "../../../../scene/scene_manager.h"
+#include "../../../../ecs/component/sprite_component.h"
+#include "../../../../ecs/component/animated_sprite_component.h"
+#include "../../../../ecs/component/text_label_component.h"
+#include "../../../../ecs/component/collider2d_component.h"
+#include "../../../../ecs/component/color_rect_component.h"
+#include "../../../../ecs/component/parallax_component.h"
+#include "../../../../ecs/component/script_component.h"
 
-// Shader Instance
-int cre_pkpy_api_shader_instance_delete(pkpy_vm* vm);
-int cre_pkpy_api_shader_instance_create_bool_param(pkpy_vm* vm);
-int cre_pkpy_api_shader_instance_set_bool_param(pkpy_vm* vm);
-int cre_pkpy_api_shader_instance_get_bool_param(pkpy_vm* vm);
-int cre_pkpy_api_shader_instance_create_int_param(pkpy_vm* vm);
-int cre_pkpy_api_shader_instance_set_int_param(pkpy_vm* vm);
-int cre_pkpy_api_shader_instance_get_int_param(pkpy_vm* vm);
-int cre_pkpy_api_shader_instance_create_float_param(pkpy_vm* vm);
-int cre_pkpy_api_shader_instance_set_float_param(pkpy_vm* vm);
-int cre_pkpy_api_shader_instance_get_float_param(pkpy_vm* vm);
-int cre_pkpy_api_shader_instance_create_float2_param(pkpy_vm* vm);
-int cre_pkpy_api_shader_instance_set_float2_param(pkpy_vm* vm);
-int cre_pkpy_api_shader_instance_get_float2_param(pkpy_vm* vm);
-int cre_pkpy_api_shader_instance_create_float3_param(pkpy_vm* vm);
-int cre_pkpy_api_shader_instance_set_float3_param(pkpy_vm* vm);
-int cre_pkpy_api_shader_instance_get_float3_param(pkpy_vm* vm);
-int cre_pkpy_api_shader_instance_create_float4_param(pkpy_vm* vm);
-int cre_pkpy_api_shader_instance_set_float4_param(pkpy_vm* vm);
-int cre_pkpy_api_shader_instance_get_float4_param(pkpy_vm* vm);
-
-// Shader Util
-int cre_pkpy_api_shader_util_compile_shader(pkpy_vm* vm);
-int cre_pkpy_api_shader_util_compile_shader_raw(pkpy_vm* vm);
-int cre_pkpy_api_shader_util_set_screen_shader(pkpy_vm* vm);
-int cre_pkpy_api_shader_util_get_current_screen_shader(pkpy_vm* vm);
-int cre_pkpy_api_shader_util_reset_screen_shader_to_default(pkpy_vm* vm);
-
-// Node
-int cre_pkpy_api_node_new(pkpy_vm* vm);
-int cre_pkpy_api_node_get_name(pkpy_vm* vm);
-int cre_pkpy_api_node_add_child(pkpy_vm* vm);
-int cre_pkpy_api_node_get_child(pkpy_vm* vm);
-int cre_pkpy_api_node_get_children(pkpy_vm* vm);
-int cre_pkpy_api_node_get_parent(pkpy_vm* vm);
-int cre_pkpy_api_node_queue_deletion(pkpy_vm* vm);
-int cre_pkpy_api_node_is_queued_for_deletion(pkpy_vm* vm);
-int cre_pkpy_api_node_set_time_dilation(pkpy_vm* vm);
-int cre_pkpy_api_node_get_time_dilation(pkpy_vm* vm);
-int cre_pkpy_api_node_get_total_time_dilation(pkpy_vm* vm);
-// TODO: These are implemented in python, may want to move later..
-//int cre_pkpy_api_node_create_event(pkpy_vm* vm);
-//int cre_pkpy_api_node_subscribe_to_event(pkpy_vm* vm);
-//int cre_pkpy_api_node_broadcast_event(pkpy_vm* vm);
-
-// Node2D
-int cre_pkpy_api_node2d_set_position(pkpy_vm* vm);
-int cre_pkpy_api_node2d_add_to_position(pkpy_vm* vm);
-int cre_pkpy_api_node2d_get_position(pkpy_vm* vm);
-int cre_pkpy_api_node2d_get_global_position(pkpy_vm* vm);
-int cre_pkpy_api_node2d_set_scale(pkpy_vm* vm);
-int cre_pkpy_api_node2d_add_to_scale(pkpy_vm* vm);
-int cre_pkpy_api_node2d_get_scale(pkpy_vm* vm);
-int cre_pkpy_api_node2d_set_rotation(pkpy_vm* vm);
-int cre_pkpy_api_node2d_add_to_rotation(pkpy_vm* vm);
-int cre_pkpy_api_node2d_get_rotation(pkpy_vm* vm);
-int cre_pkpy_api_node2d_set_z_index(pkpy_vm* vm);
-int cre_pkpy_api_node2d_get_z_index(pkpy_vm* vm);
-int cre_pkpy_api_node2d_set_z_index_relative_to_parent(pkpy_vm* vm);
-int cre_pkpy_api_node2d_get_z_index_relative_to_parent(pkpy_vm* vm);
-int cre_pkpy_api_node2d_set_ignore_camera(pkpy_vm* vm);
-int cre_pkpy_api_node2d_get_ignore_camera(pkpy_vm* vm);
-
-// Sprite
-int cre_pkpy_api_sprite_get_texture(pkpy_vm* vm);
-int cre_pkpy_api_sprite_set_texture(pkpy_vm* vm);
-int cre_pkpy_api_sprite_get_draw_source(pkpy_vm* vm);
-int cre_pkpy_api_sprite_set_draw_source(pkpy_vm* vm);
-int cre_pkpy_api_sprite_get_flip_h(pkpy_vm* vm);
-int cre_pkpy_api_sprite_set_flip_h(pkpy_vm* vm);
-int cre_pkpy_api_sprite_get_flip_v(pkpy_vm* vm);
-int cre_pkpy_api_sprite_set_flip_v(pkpy_vm* vm);
-int cre_pkpy_api_sprite_get_modulate(pkpy_vm* vm);
-int cre_pkpy_api_sprite_set_modulate(pkpy_vm* vm);
-int cre_pkpy_api_sprite_get_origin(pkpy_vm* vm);
-int cre_pkpy_api_sprite_set_origin(pkpy_vm* vm);
-int cre_pkpy_api_sprite_get_shader_instance(pkpy_vm* vm);
-int cre_pkpy_api_sprite_set_shader_instance(pkpy_vm* vm);
-
-// Animated Sprite
-int cre_pkpy_api_animated_sprite_play(pkpy_vm* vm);
-int cre_pkpy_api_animated_sprite_stop(pkpy_vm* vm);
-int cre_pkpy_api_animated_sprite_set_current_animation_frame(pkpy_vm* vm);
-int cre_pkpy_api_animated_sprite_add_animation(pkpy_vm* vm);
-int cre_pkpy_api_animated_sprite_get_stagger_animation_start_times(pkpy_vm* vm);
-int cre_pkpy_api_animated_sprite_set_stagger_animation_start_times(pkpy_vm* vm);
-int cre_pkpy_api_animated_sprite_get_flip_h(pkpy_vm* vm);
-int cre_pkpy_api_animated_sprite_set_flip_h(pkpy_vm* vm);
-int cre_pkpy_api_animated_sprite_get_flip_v(pkpy_vm* vm);
-int cre_pkpy_api_animated_sprite_set_flip_v(pkpy_vm* vm);
-int cre_pkpy_api_animated_sprite_get_modulate(pkpy_vm* vm);
-int cre_pkpy_api_animated_sprite_set_modulate(pkpy_vm* vm);
-int cre_pkpy_api_animated_sprite_get_origin(pkpy_vm* vm);
-int cre_pkpy_api_animated_sprite_set_origin(pkpy_vm* vm);
-int cre_pkpy_api_animated_sprite_get_shader_instance(pkpy_vm* vm);
-int cre_pkpy_api_animated_sprite_set_shader_instance(pkpy_vm* vm);
-
-// Text Label
-int cre_pkpy_api_text_label_get_text(pkpy_vm* vm);
-int cre_pkpy_api_text_label_set_text(pkpy_vm* vm);
-int cre_pkpy_api_text_label_get_color(pkpy_vm* vm);
-int cre_pkpy_api_text_label_set_color(pkpy_vm* vm);
-int cre_pkpy_api_text_label_set_font_uid(pkpy_vm* vm);
-
-// Collider2D
-int cre_pkpy_api_collider2d_get_extents(pkpy_vm* vm);
-int cre_pkpy_api_collider2d_set_extents(pkpy_vm* vm);
-int cre_pkpy_api_collider2d_get_color(pkpy_vm* vm);
-int cre_pkpy_api_collider2d_set_color(pkpy_vm* vm);
-
-// Color Rect
-int cre_pkpy_api_color_rect_get_size(pkpy_vm* vm);
-int cre_pkpy_api_color_rect_set_size(pkpy_vm* vm);
-int cre_pkpy_api_color_rect_get_color(pkpy_vm* vm);
-int cre_pkpy_api_color_rect_set_color(pkpy_vm* vm);
-
-// Parallax
-int cre_pkpy_api_parallax_get_scroll_speed(pkpy_vm* vm);
-int cre_pkpy_api_parallax_set_scroll_speed(pkpy_vm* vm);
-
-// Scene Tree
-int cre_pkpy_api_scene_tree_change_scene(pkpy_vm* vm);
-int cre_pkpy_api_scene_tree_get_root(pkpy_vm* vm);
-
-// Scene Manager
-int cre_pkpy_api_scene_manager_process_queued_creation_entities(pkpy_vm* vm);
-int cre_pkpy_api_scene_manager_process_queued_scene_change(pkpy_vm* vm);
-
-void cre_pkpy_api_load_internal_modules(pkpy_vm* vm) {
-    // Load internal first
-    cre_pkpy_util_create_module(vm, &(CrePPModule) {
-        .name = "crescent_internal",
-        .functions = {
-            // Shader Instance
-            {.signature = "shader_instance_delete(shader_id: int) -> bool", .function = cre_pkpy_api_shader_instance_delete},
-            {.signature = "shader_instance_create_bool_param(shader_id: int, initial_value: bool) -> None", .function = cre_pkpy_api_shader_instance_create_bool_param},
-            {.signature = "shader_instance_set_bool_param(shader_id: int, name: str, value: bool) -> None", .function = cre_pkpy_api_shader_instance_set_bool_param},
-            {.signature = "shader_instance_get_bool_param(shader_id: int, name: str) -> bool", .function = cre_pkpy_api_shader_instance_get_bool_param},
-            {.signature = "shader_instance_create_int_param(shader_id: int, initial_value: int) -> None", .function = cre_pkpy_api_shader_instance_create_int_param},
-            {.signature = "shader_instance_set_int_param(shader_id: int, name: str, value: int) -> None", .function = cre_pkpy_api_shader_instance_set_int_param},
-            {.signature = "shader_instance_get_int_param(shader_id: int, name: str) -> int", .function = cre_pkpy_api_shader_instance_get_int_param},
-            {.signature = "shader_instance_create_float_param(shader_id: int, initial_value: float) -> None", .function = cre_pkpy_api_shader_instance_create_float_param},
-            {.signature = "shader_instance_set_float_param(shader_id: int, name: str, value: float) -> None", .function = cre_pkpy_api_shader_instance_set_float_param},
-            {.signature = "shader_instance_get_float_param(shader_id: int, name: str) -> float", .function = cre_pkpy_api_shader_instance_get_float_param},
-            {.signature = "shader_instance_create_float2_param(shader_id: int, initial_value_x: float, initial_value_y: float) -> None", .function = cre_pkpy_api_shader_instance_create_float2_param},
-            {.signature = "shader_instance_set_float2_param(shader_id: int, name: str, value_x: float, value_y: float) -> None", .function = cre_pkpy_api_shader_instance_set_float2_param},
-            {.signature = "shader_instance_get_float2_param(shader_id: int, name: str) -> Tuple[float, float]", .function = cre_pkpy_api_shader_instance_get_float2_param},
-            {.signature = "shader_instance_create_float3_param(shader_id: int, initial_value_x: float, initial_value_y: float, initial_value_z: float) -> None", .function = cre_pkpy_api_shader_instance_create_float3_param},
-            {.signature = "shader_instance_set_float3_param(shader_id: int, name: str, value_x: float, value_y: float, value_z: float) -> None", .function = cre_pkpy_api_shader_instance_set_float3_param},
-            {.signature = "shader_instance_get_float3_param(shader_id: int, name: str) -> Tuple[float, float, float]", .function = cre_pkpy_api_shader_instance_get_float3_param},
-            {.signature = "shader_instance_create_float4_param(shader_id: int, initial_value_x: float, initial_value_y: float, initial_value_z: float, initial_value_w: float) -> None", .function = cre_pkpy_api_shader_instance_create_float4_param},
-            {.signature = "shader_instance_set_float4_param(shader_id: int, name: str, value_x: float, value_y: float, value_z: float, value_w: float) -> None", .function = cre_pkpy_api_shader_instance_set_float4_param},
-            {.signature = "shader_instance_get_float4_param(shader_id: int, name: str) -> Tuple[float, float, float, float]", .function = cre_pkpy_api_shader_instance_get_float4_param},
-            // Shader Util
-            {.signature = "shader_util_compile_shader(shader_path: str) -> int", .function = cre_pkpy_api_shader_util_compile_shader},
-            {.signature = "shader_util_compile_shader_raw(vertex_path, fragment_path) -> int", .function = cre_pkpy_api_shader_util_compile_shader_raw},
-            {.signature = "shader_util_set_screen_shader(shader_path: str) -> bool", .function = cre_pkpy_api_shader_util_set_screen_shader},
-            {.signature = "shader_util_get_current_screen_shader() -> int", .function = cre_pkpy_api_shader_util_get_current_screen_shader},
-            {.signature = "shader_util_reset_screen_shader_to_default() -> None", .function = cre_pkpy_api_shader_util_reset_screen_shader_to_default},
-            // Node
-            {.signature = "node_new(class_path: str, class_name: str, node_type_flag: int) -> \"Node\"", .function = cre_pkpy_api_node_new},
-            {.signature = "node_get_name(entity_id: int) -> str", .function = cre_pkpy_api_node_get_name},
-            {.signature = "node_add_child(parent_entity_id: int, child_entity_id: int) -> None", .function = cre_pkpy_api_node_add_child},
-            {.signature = "node_get_child(parent_entity_id: int, child_entity_name: str) -> Optional[\"Node\"]", .function = cre_pkpy_api_node_get_child},
-            {.signature = "node_get_children(entity_id: int) -> Tuple[\"Node\", ...]", .function = cre_pkpy_api_node_get_children},
-            {.signature = "node_get_parent(child_entity_id: int) -> Optional[\"Node\"]", .function = cre_pkpy_api_node_get_parent},
-            {.signature = "node_queue_deletion(entity_id: int) -> None", .function = cre_pkpy_api_node_queue_deletion},
-            {.signature = "node_is_queued_for_deletion(entity_id: int) -> bool", .function = cre_pkpy_api_node_is_queued_for_deletion},
-            {.signature = "node_set_time_dilation(entity_id: int, dilation: float) -> None", .function = cre_pkpy_api_node_set_time_dilation},
-            {.signature = "node_get_time_dilation(entity_id: int) -> float", .function = cre_pkpy_api_node_get_time_dilation},
-            {.signature = "node_get_total_time_dilation(entity_id: int) -> float", .function = cre_pkpy_api_node_get_total_time_dilation},
-            // Node2D
-            {.signature = "node2d_set_position(entity_id: int, x: float, y: float) -> None", .function = cre_pkpy_api_node2d_set_position},
-            {.signature = "node2d_add_to_position(entity_id: int, x: float, y: float) -> None", .function = cre_pkpy_api_node2d_add_to_position},
-            {.signature = "node2d_get_position(entity_id: int) -> Tuple[float, float]", .function = cre_pkpy_api_node2d_get_position},
-            {.signature = "node2d_get_global_position(entity_id: int) -> Tuple[float, float]", .function = cre_pkpy_api_node2d_get_global_position},
-            {.signature = "node2d_set_scale(entity_id: int, x: float, y: float) -> None", .function = cre_pkpy_api_node2d_set_scale},
-            {.signature = "node2d_add_to_scale(entity_id: int, x: float, y: float) -> None", .function = cre_pkpy_api_node2d_add_to_scale},
-            {.signature = "node2d_get_scale(entity_id: int) -> Tuple[float, float]", .function = cre_pkpy_api_node2d_get_scale},
-            {.signature = "node2d_set_rotation(entity_id: int, rotation: float) -> None", .function = cre_pkpy_api_node2d_set_rotation},
-            {.signature = "node2d_add_to_rotation(entity_id: int, rotation: float) -> None", .function = cre_pkpy_api_node2d_add_to_rotation},
-            {.signature = "node2d_get_rotation(entity_id: int) -> float", .function = cre_pkpy_api_node2d_get_rotation},
-            {.signature = "node2d_set_z_index(entity_id: int, z_index: int) -> None", .function = cre_pkpy_api_node2d_set_z_index},
-            {.signature = "node2d_get_z_index(entity_id: int) -> int", .function = cre_pkpy_api_node2d_get_z_index},
-            {.signature = "node2d_set_z_index_relative_to_parent(entity_id: int, is_relative: bool) -> None", .function = cre_pkpy_api_node2d_set_z_index_relative_to_parent},
-            {.signature = "node2d_get_z_index_relative_to_parent(entity_id: int) -> bool", .function = cre_pkpy_api_node2d_get_z_index_relative_to_parent},
-            {.signature = "node2d_set_ignore_camera(entity_id: int, ignore_camera: bool) -> None", .function = cre_pkpy_api_node2d_set_ignore_camera},
-            {.signature = "node2d_get_ignore_camera(entity_id: int) -> bool", .function = cre_pkpy_api_node2d_get_ignore_camera},
-            // Sprite
-            {.signature = "sprite_get_texture(entity_id: int) -> Tuple[str, str, str, bool]", .function = cre_pkpy_api_sprite_get_texture},
-            {.signature = "sprite_set_texture(entity_id: int, file_path: str) -> None", .function = cre_pkpy_api_sprite_set_texture},
-            {.signature = "sprite_get_draw_source(entity_id: int) -> Tuple[float, float, float, float]", .function = cre_pkpy_api_sprite_get_draw_source},
-            {.signature = "sprite_set_draw_source(entity_id: int, x: float, y: float, w: float, h: float) -> None", .function = cre_pkpy_api_sprite_set_draw_source},
-            {.signature = "sprite_get_flip_h(entity_id: int) -> bool", .function = cre_pkpy_api_sprite_get_flip_h},
-            {.signature = "sprite_set_flip_h(entity_id: int, flip_h: bool) -> None", .function = cre_pkpy_api_sprite_set_flip_h},
-            {.signature = "sprite_get_flip_v(entity_id: int) -> bool", .function = cre_pkpy_api_sprite_get_flip_v},
-            {.signature = "sprite_set_flip_v(entity_id: int, flip_v: bool) -> None", .function = cre_pkpy_api_sprite_set_flip_v},
-            {.signature = "sprite_get_modulate(entity_id: int) -> Tuple[int, int, int, int]", .function = cre_pkpy_api_sprite_get_modulate},
-            {.signature = "sprite_set_modulate(entity_id: int, r: int, g: int, b: int, a: int) -> None", .function = cre_pkpy_api_sprite_set_modulate},
-            {.signature = "sprite_get_origin(entity_id: int) -> Tuple[float, float]", .function = cre_pkpy_api_sprite_get_origin},
-            {.signature = "sprite_set_origin(entity_id: int, x: float, y: float) -> None", .function = cre_pkpy_api_sprite_set_origin},
-            {.signature = "sprite_get_shader_instance(entity_id: int) -> int", .function = cre_pkpy_api_sprite_get_shader_instance},
-            {.signature = "sprite_set_shader_instance(entity_id: int, shader_instance_id: int) -> None", .function = cre_pkpy_api_sprite_set_shader_instance},
-            // Animated Sprite
-            {.signature = "animated_sprite_play(entity_id: int, animation_name: str) -> bool", .function = cre_pkpy_api_animated_sprite_play},
-            {.signature = "animated_sprite_stop(entity_id: int) -> None", .function = cre_pkpy_api_animated_sprite_stop},
-            {.signature = "animated_sprite_set_current_animation_frame(entity_id: int, frame: int) -> None", .function = cre_pkpy_api_animated_sprite_set_current_animation_frame},
-            {.signature = "animated_sprite_add_animation(entity_id: int, name: str, speed: float, does_loop: bool, frame_count: int, *args) -> None", .function = cre_pkpy_api_animated_sprite_add_animation},
-            {.signature = "animated_sprite_get_stagger_animation_start_times(entity_id: int) -> bool", .function = cre_pkpy_api_animated_sprite_get_stagger_animation_start_times},
-            {.signature = "animated_sprite_set_stagger_animation_start_times(entity_id: int, stagger: bool) -> None", .function = cre_pkpy_api_animated_sprite_set_stagger_animation_start_times},
-            {.signature = "animated_sprite_get_flip_h(entity_id: int) -> bool", .function = cre_pkpy_api_animated_sprite_get_flip_h},
-            {.signature = "animated_sprite_set_flip_h(entity_id: int, flip_h: bool) -> None", .function = cre_pkpy_api_animated_sprite_set_flip_h},
-            {.signature = "animated_sprite_get_flip_v(entity_id: int) -> bool", .function = cre_pkpy_api_animated_sprite_get_flip_v},
-            {.signature = "animated_sprite_set_flip_v(entity_id: int, flip_v: bool) -> None", .function = cre_pkpy_api_animated_sprite_set_flip_v},
-            {.signature = "animated_sprite_get_modulate(entity_id: int) -> Tuple[int, int, int, int]", .function = cre_pkpy_api_animated_sprite_get_modulate},
-            {.signature = "animated_sprite_set_modulate(entity_id: int, r: int, g: int, b: int, a: int) -> None", .function = cre_pkpy_api_animated_sprite_set_modulate},
-            {.signature = "animated_sprite_get_origin(entity_id: int) -> Tuple[float, float]", .function = cre_pkpy_api_animated_sprite_get_origin},
-            {.signature = "animated_sprite_set_origin(entity_id: int, x: float, y: float) -> None", .function = cre_pkpy_api_animated_sprite_set_origin},
-            {.signature = "animated_sprite_get_shader_instance(entity_id: int) -> int", .function = cre_pkpy_api_animated_sprite_get_shader_instance},
-            {.signature = "animated_sprite_set_shader_instance(entity_id: int, shader_instance_id: int) -> None", .function = cre_pkpy_api_animated_sprite_set_shader_instance},
-            // Text Label
-            {.signature = "text_label_get_text(entity_id: int) -> str", .function = cre_pkpy_api_text_label_get_text},
-            {.signature = "text_label_set_text(entity_id: int, text: str) -> None", .function = cre_pkpy_api_text_label_set_text},
-            {.signature = "text_label_get_color(entity_id: int) -> Tuple[int, int, int, int]", .function = cre_pkpy_api_text_label_get_color},
-            {.signature = "text_label_set_color(entity_id: int, r: int, g: int, b: int, a: int) -> None", .function = cre_pkpy_api_text_label_set_color},
-            {.signature = "text_label_set_font_uid(entity_id: int, uid: str) -> None", .function = cre_pkpy_api_text_label_set_font_uid},
-            // Collider2D
-            {.signature = "collider2d_get_extents(entity_id: int) -> Tuple[float, float]", .function = cre_pkpy_api_collider2d_get_extents},
-            {.signature = "collider2d_set_extents(entity_id: int, w: float, h: float)", .function = cre_pkpy_api_collider2d_set_extents},
-            {.signature = "collider2d_get_color(entity_id: int) -> Tuple[int, int, int, int]", .function = cre_pkpy_api_collider2d_get_color},
-            {.signature = "collider2d_set_color(entity_id: int, r: int, g: int, b: int, a: int) -> None", .function = cre_pkpy_api_collider2d_set_color},
-            // Color Rect
-            {.signature = "color_rect_get_size(entity_id: int) -> Tuple[float, float]", .function = cre_pkpy_api_color_rect_get_size},
-            {.signature = "color_rect_set_size(entity_id: int, w: float, h: float) -> None", .function = cre_pkpy_api_color_rect_set_size},
-            {.signature = "color_rect_get_color(entity_id: int) -> Tuple[int, int, int, int]", .function = cre_pkpy_api_color_rect_get_color},
-            {.signature = "color_rect_set_color(entity_id: int, r: int, g: int, b: int, a: int) -> None", .function = cre_pkpy_api_color_rect_set_color},
-            // Parallax
-            {.signature = "parallax_get_scroll_speed(entity_id: int) -> Tuple[float, float]", .function = cre_pkpy_api_parallax_get_scroll_speed},
-            {.signature = "parallax_set_scroll_speed(entity_id: int, x: float, y: float) -> None", .function = cre_pkpy_api_parallax_set_scroll_speed},
-            // Scene Tree
-            {.signature = "scene_tree_change_scene(path: str) -> None", .function = cre_pkpy_api_scene_tree_change_scene},
-            {.signature = "scene_tree_get_root()", .function = cre_pkpy_api_scene_tree_get_root},
-            // Scene Manager
-            {.signature = "_scene_manager_process_queued_creation_entities() -> None", .function = cre_pkpy_api_scene_manager_process_queued_creation_entities},
-            {.signature = "_scene_manager_process_queued_scene_change() -> None", .function = cre_pkpy_api_scene_manager_process_queued_scene_change},
-
-            { NULL, NULL },
-        }
-    });
-
-    // Now load front facing api
-    cre_pkpy_util_create_from_string(vm, CRE_PKPY_MODULE_NAME_CRESCENT, CRE_PKPY_CRESCENT_SOURCE);
-}
-
-//--- SHADER INSTANCE ---//
-
-int cre_pkpy_api_shader_instance_delete(pkpy_vm* vm) {
-    int pyShaderId;
-    pkpy_to_int(vm, 0, &pyShaderId);
-
-    const SEShaderInstanceId shaderId = (SEShaderInstanceId)pyShaderId;
-    SEShaderInstance* shaderInstance = se_shader_cache_get_instance(shaderId);
-    bool hasDeletedInstance = false;
-    if (shaderInstance != NULL) {
-        se_shader_cache_remove_instance(shaderId);
-        se_shader_instance_destroy(shaderInstance);
-        hasDeletedInstance = true;
-    }
-    pkpy_push_bool(vm, hasDeletedInstance);
-    return 1;
-}
-
-int cre_pkpy_api_shader_instance_create_bool_param(pkpy_vm* vm) {
-    int pyShaderId;
-    pkpy_CString pyParamName;
-    bool value;
-    pkpy_to_int(vm, 0, &pyShaderId);
-    pkpy_to_string(vm, 1, &pyParamName);
-    pkpy_to_bool(vm, 2, &value);
-
-    const SEShaderInstanceId shaderId = (SEShaderInstanceId)pyShaderId;
-    SEShaderInstance* shaderInstance = se_shader_cache_get_instance(shaderId);
-    se_shader_instance_param_create_bool(shaderInstance, pyParamName.data, value);
-    return 0;
-}
-
-int cre_pkpy_api_shader_instance_set_bool_param(pkpy_vm* vm) {
-    int pyShaderId;
-    pkpy_CString pyParamName;
-    bool value;
-    pkpy_to_int(vm, 0, &pyShaderId);
-    pkpy_to_string(vm, 1, &pyParamName);
-    pkpy_to_bool(vm, 2, &value);
-
-    const SEShaderInstanceId shaderId = (SEShaderInstanceId)pyShaderId;
-    SEShaderInstance* shaderInstance = se_shader_cache_get_instance(shaderId);
-    se_shader_instance_param_update_bool(shaderInstance, pyParamName.data, value);
-    return 0;
-}
-
-int cre_pkpy_api_shader_instance_get_bool_param(pkpy_vm* vm) {
-    int pyShaderId;
-    pkpy_CString pyParamName;
-    pkpy_to_int(vm, 0, &pyShaderId);
-    pkpy_to_string(vm, 1, &pyParamName);
-
-    const SEShaderInstanceId shaderId = (SEShaderInstanceId)pyShaderId;
-    SEShaderInstance* shaderInstance = se_shader_cache_get_instance(shaderId);
-    const bool paramValue = se_shader_instance_param_get_bool(shaderInstance, pyParamName.data);
-    pkpy_push_bool(vm, paramValue);
-    return 1;
-}
-
-int cre_pkpy_api_shader_instance_create_int_param(pkpy_vm* vm) {
-    int pyShaderId;
-    pkpy_CString pyParamName;
-    int value;
-    pkpy_to_int(vm, 0, &pyShaderId);
-    pkpy_to_string(vm, 1, &pyParamName);
-    pkpy_to_int(vm, 2, &value);
-
-    const SEShaderInstanceId shaderId = (SEShaderInstanceId)pyShaderId;
-    SEShaderInstance* shaderInstance = se_shader_cache_get_instance(shaderId);
-    se_shader_instance_param_create_int(shaderInstance, pyParamName.data, value);
-    return 0;
-}
-
-int cre_pkpy_api_shader_instance_set_int_param(pkpy_vm* vm) {
-    int pyShaderId;
-    pkpy_CString pyParamName;
-    int value;
-    pkpy_to_int(vm, 0, &pyShaderId);
-    pkpy_to_string(vm, 1, &pyParamName);
-    pkpy_to_int(vm, 2, &value);
-
-    const SEShaderInstanceId shaderId = (SEShaderInstanceId)pyShaderId;
-    SEShaderInstance* shaderInstance = se_shader_cache_get_instance(shaderId);
-    se_shader_instance_param_update_int(shaderInstance, pyParamName.data, value);
-    return 0;
-}
-
-int cre_pkpy_api_shader_instance_get_int_param(pkpy_vm* vm) {
-    int pyShaderId;
-    pkpy_CString pyParamName;
-    pkpy_to_int(vm, 0, &pyShaderId);
-    pkpy_to_string(vm, 1, &pyParamName);
-
-    const SEShaderInstanceId shaderId = (SEShaderInstanceId)pyShaderId;
-    SEShaderInstance* shaderInstance = se_shader_cache_get_instance(shaderId);
-    const int paramValue = se_shader_instance_param_get_int(shaderInstance, pyParamName.data);
-    pkpy_push_int(vm, paramValue);
-    return 1;
-}
-
-int cre_pkpy_api_shader_instance_create_float_param(pkpy_vm* vm) {
-    int pyShaderId;
-    pkpy_CString pyParamName;
-    double value;
-    pkpy_to_int(vm, 0, &pyShaderId);
-    pkpy_to_string(vm, 1, &pyParamName);
-    pkpy_to_float(vm, 2, &value);
-
-    const SEShaderInstanceId shaderId = (SEShaderInstanceId)pyShaderId;
-    SEShaderInstance* shaderInstance = se_shader_cache_get_instance(shaderId);
-    se_shader_instance_param_create_float(shaderInstance, pyParamName.data, (float)value);
-    return 0;
-}
-
-int cre_pkpy_api_shader_instance_set_float_param(pkpy_vm* vm) {
-    int pyShaderId;
-    pkpy_CString pyParamName;
-    double value;
-    pkpy_to_int(vm, 0, &pyShaderId);
-    pkpy_to_string(vm, 1, &pyParamName);
-    pkpy_to_float(vm, 2, &value);
-
-    const SEShaderInstanceId shaderId = (SEShaderInstanceId)pyShaderId;
-    SEShaderInstance* shaderInstance = se_shader_cache_get_instance(shaderId);
-    se_shader_instance_param_update_float(shaderInstance, pyParamName.data, (float)value);
-    return 0;
-}
-
-int cre_pkpy_api_shader_instance_get_float_param(pkpy_vm* vm) {
-    int pyShaderId;
-    pkpy_CString pyParamName;
-    pkpy_to_int(vm, 0, &pyShaderId);
-    pkpy_to_string(vm, 1, &pyParamName);
-
-    const SEShaderInstanceId shaderId = (SEShaderInstanceId)pyShaderId;
-    SEShaderInstance* shaderInstance = se_shader_cache_get_instance(shaderId);
-    const float paramValue = se_shader_instance_param_get_float(shaderInstance, pyParamName.data);
-    pkpy_push_float(vm, paramValue);
-    return 1;
-}
-
-int cre_pkpy_api_shader_instance_create_float2_param(pkpy_vm* vm) {
-    int pyShaderId;
-    pkpy_CString pyParamName;
-    double valueX;
-    double valueY;
-    pkpy_to_int(vm, 0, &pyShaderId);
-    pkpy_to_string(vm, 1, &pyParamName);
-    pkpy_to_float(vm, 2, &valueX);
-    pkpy_to_float(vm, 3, &valueY);
-
-    const SEShaderInstanceId shaderId = (SEShaderInstanceId)pyShaderId;
-    SEShaderInstance* shaderInstance = se_shader_cache_get_instance(shaderId);
-    se_shader_instance_param_create_float2(shaderInstance, pyParamName.data, (SEVector2){ (float)valueX, (float)valueY });
-    return 0;
-}
-
-int cre_pkpy_api_shader_instance_set_float2_param(pkpy_vm* vm) {
-    int pyShaderId;
-    pkpy_CString pyParamName;
-    double valueX;
-    double valueY;
-    pkpy_to_int(vm, 0, &pyShaderId);
-    pkpy_to_string(vm, 1, &pyParamName);
-    pkpy_to_float(vm, 2, &valueX);
-    pkpy_to_float(vm, 3, &valueY);
-
-    const SEShaderInstanceId shaderId = (SEShaderInstanceId)pyShaderId;
-    SEShaderInstance* shaderInstance = se_shader_cache_get_instance(shaderId);
-    se_shader_instance_param_update_float2(shaderInstance, pyParamName.data, (SEVector2){ (float)valueX, (float)valueY });
-    return 0;
-}
-
-int cre_pkpy_api_shader_instance_get_float2_param(pkpy_vm* vm) {
-    int pyShaderId;
-    pkpy_CString pyParamName;
-    pkpy_to_int(vm, 0, &pyShaderId);
-    pkpy_to_string(vm, 1, &pyParamName);
-
-    const SEShaderInstanceId shaderId = (SEShaderInstanceId)pyShaderId;
-    SEShaderInstance* shaderInstance = se_shader_cache_get_instance(shaderId);
-    const SEVector2 paramValue = se_shader_instance_param_get_float2(shaderInstance, pyParamName.data);
-    pkpy_push_float(vm, paramValue.x);
-    pkpy_push_float(vm, paramValue.y);
-    return 2;
-}
-
-int cre_pkpy_api_shader_instance_create_float3_param(pkpy_vm* vm) {
-    int pyShaderId;
-    pkpy_CString pyParamName;
-    double valueX;
-    double valueY;
-    double valueZ;
-    pkpy_to_int(vm, 0, &pyShaderId);
-    pkpy_to_string(vm, 1, &pyParamName);
-    pkpy_to_float(vm, 2, &valueX);
-    pkpy_to_float(vm, 3, &valueY);
-    pkpy_to_float(vm, 4, &valueZ);
-
-    const SEShaderInstanceId shaderId = (SEShaderInstanceId)pyShaderId;
-    SEShaderInstance* shaderInstance = se_shader_cache_get_instance(shaderId);
-    se_shader_instance_param_create_float3(shaderInstance, pyParamName.data, (SEVector3){ (float)valueX, (float)valueY, (float)valueZ });
-    return 0;
-}
-
-int cre_pkpy_api_shader_instance_set_float3_param(pkpy_vm* vm) {
-    int pyShaderId;
-    pkpy_CString pyParamName;
-    double valueX;
-    double valueY;
-    double valueZ;
-    pkpy_to_int(vm, 0, &pyShaderId);
-    pkpy_to_string(vm, 1, &pyParamName);
-    pkpy_to_float(vm, 2, &valueX);
-    pkpy_to_float(vm, 3, &valueY);
-    pkpy_to_float(vm, 4, &valueZ);
-
-    const SEShaderInstanceId shaderId = (SEShaderInstanceId)pyShaderId;
-    SEShaderInstance* shaderInstance = se_shader_cache_get_instance(shaderId);
-    se_shader_instance_param_update_float3(shaderInstance, pyParamName.data, (SEVector3){ (float)valueX, (float)valueY, (float)valueZ });
-    return 0;
-}
-
-int cre_pkpy_api_shader_instance_get_float3_param(pkpy_vm* vm) {
-    int pyShaderId;
-    pkpy_CString pyParamName;
-    pkpy_to_int(vm, 0, &pyShaderId);
-    pkpy_to_string(vm, 1, &pyParamName);
-
-    const SEShaderInstanceId shaderId = (SEShaderInstanceId)pyShaderId;
-    SEShaderInstance* shaderInstance = se_shader_cache_get_instance(shaderId);
-    const SEVector3 paramValue = se_shader_instance_param_get_float3(shaderInstance, pyParamName.data);
-    pkpy_push_float(vm, paramValue.x);
-    pkpy_push_float(vm, paramValue.y);
-    pkpy_push_float(vm, paramValue.z);
-    return 3;
-}
-
-int cre_pkpy_api_shader_instance_create_float4_param(pkpy_vm* vm) {
-    int pyShaderId;
-    pkpy_CString pyParamName;
-    double valueX;
-    double valueY;
-    double valueZ;
-    double valueW;
-    pkpy_to_int(vm, 0, &pyShaderId);
-    pkpy_to_string(vm, 1, &pyParamName);
-    pkpy_to_float(vm, 2, &valueX);
-    pkpy_to_float(vm, 3, &valueY);
-    pkpy_to_float(vm, 4, &valueZ);
-    pkpy_to_float(vm, 5, &valueW);
-
-    const SEShaderInstanceId shaderId = (SEShaderInstanceId)pyShaderId;
-    SEShaderInstance* shaderInstance = se_shader_cache_get_instance(shaderId);
-    se_shader_instance_param_create_float4(shaderInstance, pyParamName.data, (SEVector4){ (float)valueX, (float)valueY, (float)valueZ, (float)valueW });
-    return 0;
-}
-
-int cre_pkpy_api_shader_instance_set_float4_param(pkpy_vm* vm) {
-    int pyShaderId;
-    pkpy_CString pyParamName;
-    double valueX;
-    double valueY;
-    double valueZ;
-    double valueW;
-    pkpy_to_int(vm, 0, &pyShaderId);
-    pkpy_to_string(vm, 1, &pyParamName);
-    pkpy_to_float(vm, 2, &valueX);
-    pkpy_to_float(vm, 3, &valueY);
-    pkpy_to_float(vm, 4, &valueZ);
-    pkpy_to_float(vm, 5, &valueW);
-
-    const SEShaderInstanceId shaderId = (SEShaderInstanceId)pyShaderId;
-    SEShaderInstance* shaderInstance = se_shader_cache_get_instance(shaderId);
-    se_shader_instance_param_update_float4(shaderInstance, pyParamName.data, (SEVector4){ (float)valueX, (float)valueY, (float)valueZ, (float)valueW });
-    return 0;
-}
-
-int cre_pkpy_api_shader_instance_get_float4_param(pkpy_vm* vm) {
-    int pyShaderId;
-    pkpy_CString pyParamName;
-    pkpy_to_int(vm, 0, &pyShaderId);
-    pkpy_to_string(vm, 1, &pyParamName);
-
-    const SEShaderInstanceId shaderId = (SEShaderInstanceId)pyShaderId;
-    SEShaderInstance* shaderInstance = se_shader_cache_get_instance(shaderId);
-    const SEVector4 paramValue = se_shader_instance_param_get_float4(shaderInstance, pyParamName.data);
-    pkpy_push_float(vm, paramValue.x);
-    pkpy_push_float(vm, paramValue.y);
-    pkpy_push_float(vm, paramValue.z);
-    pkpy_push_float(vm, paramValue.w);
-    return 4;
-}
-
-//--- SHADER UTIL ---//
-
-int cre_pkpy_api_shader_util_compile_shader(pkpy_vm* vm) {
-    pkpy_CString pyShaderPath;
-    pkpy_to_string(vm, 0, &pyShaderPath);
-
-    const char* shaderPath = pyShaderPath.data;
-    const SEShaderInstanceId newId = se_shader_cache_create_instance_and_add(shaderPath);
-    SE_ASSERT_FMT(newId != SE_SHADER_INSTANCE_INVALID_ID, "Invalid shader id reading from path '%s'", shaderPath);
-    pkpy_push_int(vm, (int)newId);
-    return 1;
-}
-
-int cre_pkpy_api_shader_util_compile_shader_raw(pkpy_vm* vm) {
-    pkpy_CString pyVertexPath;
-    pkpy_CString pyFragmentPath;
-    pkpy_to_string(vm, 0, &pyVertexPath);
-    pkpy_to_string(vm, 1, &pyFragmentPath);
-
-    const char* vertexPath = pyVertexPath.data;
-    const char* fragmentPath = pyFragmentPath.data;
-    const SEShaderInstanceId newId = se_shader_cache_create_instance_and_add_from_raw(vertexPath, fragmentPath);
-    SE_ASSERT_FMT(newId != SE_SHADER_INSTANCE_INVALID_ID, "Invalid shader id reading from paths: vertex = '%s', fragment = '%s'", vertexPath, fragmentPath);
-    pkpy_push_int(vm, (int)newId);
-    return 1;
-}
-
-int cre_pkpy_api_shader_util_set_screen_shader(pkpy_vm* vm) {
-    int pyShaderId;
-    pkpy_to_int(vm, 0, &pyShaderId);
-
-    const SEShaderInstanceId shaderId = (SEShaderInstanceId)pyShaderId;
-    SEShaderInstance* shaderInstance = se_shader_cache_get_instance(shaderId);
-    bool hasSetShaderInstance = false;
-    if (shaderInstance) {
-        se_frame_buffer_set_screen_shader(shaderInstance);
-        hasSetShaderInstance = true;
-    }
-    pkpy_push_bool(vm, hasSetShaderInstance);
-    return 1;
-}
-
-int cre_pkpy_api_shader_util_get_current_screen_shader(pkpy_vm* vm) {
-    const SEShaderInstanceId currentScreenShaderId = 0; // TODO: Keep track of current screen shader somewhere
-    pkpy_push_int(vm, (int)currentScreenShaderId);
-    return 1;
-}
-
-int cre_pkpy_api_shader_util_reset_screen_shader_to_default(pkpy_vm* vm) {
-    se_frame_buffer_reset_to_default_screen_shader();
-    return 0;
-}
-
-//--- NODE ---//
 
 void set_node_component_from_type(CreEntity entity, const char* classPath, const char* className, NodeBaseType baseType) {
 
@@ -658,6 +52,8 @@ void set_node_component_from_type(CreEntity entity, const char* classPath, const
         cre_component_manager_set_component(entity, CreComponentDataIndex_PARALLAX, parallax_component_create());
     }
 }
+
+//--- NODE ---//
 
 int cre_pkpy_api_node_new(pkpy_vm* vm) {
     pkpy_CString pyClassPath;
@@ -923,8 +319,8 @@ int cre_pkpy_api_node2d_add_to_scale(pkpy_vm* vm) {
     const CreEntity entity = (CreEntity)pyEntityId;
     const Transform2DComponent* transformComp = (Transform2DComponent*) cre_component_manager_get_component(entity, CreComponentDataIndex_TRANSFORM_2D);
     cre_pkpy_update_entity_local_scale(entity, &(SEVector2){
-        transformComp->localTransform.scale.x + (float)scaleX,
-        transformComp->localTransform.scale.y + (float)scaleY
+            transformComp->localTransform.scale.x + (float)scaleX,
+            transformComp->localTransform.scale.y + (float)scaleY
     });
     return 0;
 }
@@ -1668,33 +1064,5 @@ int cre_pkpy_api_parallax_set_scroll_speed(pkpy_vm* vm) {
     const CreEntity entity = (CreEntity)pyEntityId;
     ParallaxComponent* parallaxComponent = (ParallaxComponent*) cre_component_manager_get_component(entity, CreComponentDataIndex_PARALLAX);
     parallaxComponent->scrollSpeed = (SEVector2){ (float)pySpeedX, (float)pySpeedY };
-    return 0;
-}
-
-//--- SCENE TREE ---//
-
-int cre_pkpy_api_scene_tree_change_scene(pkpy_vm* vm) {
-    pkpy_CString scenePath;
-    pkpy_to_string(vm, 0, &scenePath);
-
-    cre_scene_manager_queue_scene_change(scenePath.data);
-    return 0;
-}
-
-int cre_pkpy_api_scene_tree_get_root(pkpy_vm* vm) {
-    SceneTreeNode* rootNode = cre_scene_manager_get_active_scene_root();
-    SE_ASSERT(rootNode != NULL);
-    cre_pkpy_entity_instance_cache_push_or_add_default_entity_instance(vm, rootNode->entity);
-    return 1;
-}
-
-//--- SCENE MANAGER ---//
-int cre_pkpy_api_scene_manager_process_queued_creation_entities(pkpy_vm* vm) {
-    cre_scene_manager_process_queued_creation_entities();
-    return 0;
-}
-
-int cre_pkpy_api_scene_manager_process_queued_scene_change(pkpy_vm* vm) {
-    cre_scene_manager_process_queued_scene_change();
     return 0;
 }
