--- engine/src/core/scripting/python/pocketpy/cre_pkpy_api.h
@@ -1,5 +1,0 @@
-#pragma once
-
-#include "pocketpy_c.h"
-
-void cre_pkpy_api_load_internal_modules(pkpy_vm* vm);
